requirements.txt
# ==================== 后端核心依赖 ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# ==================== LangChain相关 ====================
langchain==0.1.0
langchain-community==0.0.10
openai>=1.12.0

# ==================== 数据处理和图表 ====================
pandas==2.1.3
matplotlib==3.8.2
numpy==1.26.2

# ==================== 工具库 ====================
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
pyyaml==6.0.1
httpx>=0.25.0
